+++ tests/src/Opulence/IO/FileSystemTest.php
@@ -8,7 +8,7 @@
  * @license   https://github.com/opulencephp/Opulence/blob/master/LICENSE.md
  */
 
-namespace Opulence\Files;
+namespace Opulence\IO;
 
 use DateTime;